+++ app-server/subprojects/bounded_context/challenge/domain/src/main/kotlin/club/staircrusher/challenge/domain/model/ChallengeRank.kt
@@ -6,7 +6,7 @@ class ChallengeRank(
     val id: String,
     val challengeId: String,
     val userId: String,
-    var contributionCount: Int,
+    var contributionCount: Long,
     var rank: Long,
     var createdAt: Instant,
     var updatedAt: Instant,
